refactor: 입력 로직 책임 분리
