+++ src/main/java/lotto/controller/LottoController.java
@@ -1,15 +1,10 @@
 package lotto.controller;
 
 import lotto.domain.LottoMachine;
-import lotto.domain.generator.LottoRandomGenerator;
 import lotto.domain.lottonumber.Lotto;
-import lotto.domain.lottonumber.Lottos;
 import lotto.domain.lottonumber.WinningNumbers;
-import lotto.domain.lottonumber.vo.LottoNumber;
 import lotto.domain.matchkind.LottoMatchKind;
-import lotto.domain.purchaseamount.TotalPurchaseAmount;
 import lotto.domain.winningresult.WinningResult;
-import lotto.dto.InputLottoDto;
 import lotto.dto.LottoMatchKindDto;
 import lotto.dto.LottoNumbersDto;
 import lotto.view.input.InputView;
@@ -27,59 +22,14 @@ public class LottoController {
     public LottoController(final InputView inputView, final OutputView outputView, final int lottoPrice) {
         this.inputView = inputView;
         this.outputView = outputView;
-        lottoMachine = initializeLottoService(lottoPrice);
-    }
-
-    private LottoMachine initializeLottoService(final int lottoPrice) {
-        final TotalPurchaseAmount.TotalPurchaseAmountBuilder totalPurchaseAmountBuilder =
-                inputTotalPurchaseAmount(lottoPrice);
-        final TotalPurchaseAmount totalPurchaseAmount = inputManualPurchaseAmount(totalPurchaseAmountBuilder).build();
-        final List<Lotto> manualLottos = inputManualLottos(totalPurchaseAmount.getCountOfManualLottoNumbers());
-        return new LottoMachine(new LottoRandomGenerator(), totalPurchaseAmount, manualLottos);
-    }
-
-    private TotalPurchaseAmount.TotalPurchaseAmountBuilder inputTotalPurchaseAmount(final int lottoPrice) {
-        try {
-            final String purchaseAmountInput = inputView.inputPurchaseAmount();
-            return new TotalPurchaseAmount.TotalPurchaseAmountBuilder()
-                    .setLottoPrice(lottoPrice)
-                    .setTotalAmount(purchaseAmountInput);
-        } catch (final Exception e) {
-            inputView.printErrorMessage(e.getMessage());
-            return inputTotalPurchaseAmount(lottoPrice);
-        }
-    }
-
-    private TotalPurchaseAmount.TotalPurchaseAmountBuilder inputManualPurchaseAmount(
-            final TotalPurchaseAmount.TotalPurchaseAmountBuilder totalPurchaseAmountBuilder) {
-        try {
-            final String manualPurchaseLottoAmount = inputView.inputManualPurchaseAmount();
-            return totalPurchaseAmountBuilder.setManualPurchaseAmount(manualPurchaseLottoAmount);
-        } catch (final Exception e) {
-            inputView.printErrorMessage(e.getMessage());
-            return inputManualPurchaseAmount(totalPurchaseAmountBuilder);
-        }
-    }
-
-    private List<Lotto> inputManualLottos(final int numberOfManualLottos) {
-        try {
-            final List<InputLottoDto> manualInputLottos = inputView.inputManualLottoNumbers(numberOfManualLottos);
-            final List<Lotto> manual = manualInputLottos.stream()
-                    .map(InputLottoDto::getNumbers)
-                    .map(Lotto::new)
-                    .collect(Collectors.toUnmodifiableList());
-            return new Lottos(manual).getLottos();
-        } catch (final Exception e) {
-            inputView.printErrorMessage(e.getMessage());
-            return inputManualLottos(numberOfManualLottos);
-        }
+        lottoMachine = new LottoMachineInitializer(inputView).initializeLottoService(lottoPrice);
     }
 
     public void run() {
         outputView.printPurchaseCount(
                 lottoMachine.getCountOfManualLottoNumbers(), lottoMachine.getCountOfAutoLottoNumbers());
         printLottoNumbersGroup();
-        final WinningNumbers winningNumbers = generateWinningNumbers();
+        final WinningNumbers winningNumbers = new WinningLottoGenerator().generateWinningNumbers(inputView);
         printResult(winningNumbers);
     }
 
@@ -94,17 +44,6 @@ public class LottoController {
                 .map(Lotto::getValues)
                 .map(LottoNumbersDto::new)
                 .collect(Collectors.toUnmodifiableList());
-    }
-
-    private WinningNumbers generateWinningNumbers() {
-        try {
-            final Lotto lastWinningNumbers = new Lotto(inputView.inputLastWeekWinningNumbers());
-            final LottoNumber bonusNumber = LottoNumber.from(inputView.inputBonusNumber());
-            return new WinningNumbers(lastWinningNumbers, bonusNumber);
-        } catch (final Exception e) {
-            inputView.printErrorMessage(e.getMessage());
-            return generateWinningNumbers();
-        }
     }
 
     private void printResult(WinningNumbers winningNumbers) {
